requirements.txt
numpy==1.26.1
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
Pillow==10.1.0
reportlab==4.0.6
